Changed regex expr from ast nodes to string refs
