+++ Cargo.toml
@@ -7,6 +7,6 @@ edition = "2021"
 
 [dependencies]
 ariadne = "0.2.0"
-chumsky = "0.9.2"
+chumsky = { git = "https://github.com/zesterer/chumsky" }
 logos = "0.12.1"
 snailquote = "0.3.1"
