+++ src/lexer.rs
@@ -1,160 +1,19 @@
-use logos::{Lexer, Logos, Skip};
-use snailquote::unescape;
+use crate::ast::{
+    str_to_cmp_op, str_to_exp_op, str_to_scope_spec, str_to_term_op, str_to_var_type,
+};
+use crate::ast::{CmpOp, ExpOp, ScopeSpecifier, TermOp, VarType};
+use logos::Logos;
 
-struct LexerInfo {
-    pub line: u32,
-}
-impl Default for LexerInfo {
-    fn default() -> Self {
-        Self { line: 1 }
-    }
-}
-
-fn count_newline(lex: &mut Lexer<Token>) -> Skip {
-    lex.extras.line += 1;
-
-    Skip
-}
-
-#[derive(Debug, PartialEq)]
-enum VarType {
-    I32,
-    I64,
-    F32,
-    F64,
-    Boolean,
-    Char,
-    String,
-}
-
-fn str_to_var_type(lex: &mut Lexer<Token>) -> VarType {
-    let type_str = lex.slice();
-    match type_str {
-        "i32" => VarType::I32,
-        "i64" => VarType::I64,
-        "f32" => VarType::F32,
-        "f64" => VarType::F64,
-        "boolean" => VarType::Boolean,
-        "char" => VarType::Char,
-        "string" => VarType::String,
-        _ => unreachable!(),
-    }
-}
-
-#[derive(Debug, PartialEq)]
-enum ScopeSpecifier {
-    Var,
-    Const,
-    Let,
-}
-
-fn str_to_scope_spec(lex: &mut Lexer<Token>) -> ScopeSpecifier {
-    let scope_str = lex.slice();
-    match scope_str {
-        "var" => ScopeSpecifier::Var,
-        "const" => ScopeSpecifier::Const,
-        "let" => ScopeSpecifier::Let,
-        _ => unreachable!(),
-    }
-}
-
-#[derive(Debug, PartialEq)]
-enum ExpOp {
-    Add,
-    Sub,
-}
-
-fn str_to_exp_op(lex: &mut Lexer<Token>) -> ExpOp {
-    let exp_op_str = lex.slice();
-    match exp_op_str {
-        "+" => ExpOp::Add,
-        "-" => ExpOp::Sub,
-        _ => unreachable!(),
-    }
-}
-
-#[derive(Debug, PartialEq)]
-enum TermOp {
-    Mul,
-    Div,
-    Mod,
-}
-
-fn str_to_term_op(lex: &mut Lexer<Token>) -> TermOp {
-    let term_op_str = lex.slice();
-    match term_op_str {
-        "*" => TermOp::Mul,
-        "/" => TermOp::Div,
-        "%" => TermOp::Mod,
-        _ => unreachable!(),
-    }
-}
-
-#[derive(Debug, PartialEq)]
-enum CmpOp {
-    Lt,
-    Lte,
-    Gt,
-    Gte,
-    Ne,
-    Eq,
-}
-
-fn str_to_cmp_op(lex: &mut Lexer<Token>) -> CmpOp {
-    let cmp_op_str = lex.slice();
-    match cmp_op_str {
-        ">" => CmpOp::Lt,
-        ">=" => CmpOp::Lte,
-        "<" => CmpOp::Gt,
-        "<=" => CmpOp::Gte,
-        "!=" => CmpOp::Ne,
-        "==" => CmpOp::Eq,
-        _ => unreachable!(),
-    }
-}
-
-#[derive(Debug, PartialEq)]
-struct Id(String);
-impl From<&str> for Id {
-    fn from(value: &str) -> Self {
-        Id(value.to_string())
-    }
-}
-
-fn str_to_id(lex: &mut Lexer<Token>) -> Id {
-    let slice = lex.slice();
-    Id(slice.to_string())
-}
-
-fn str_to_string(lex: &mut Lexer<Token>) -> Option<String> {
-    let slice = lex.slice();
-
-    unescape(slice).ok()
-}
-
-fn str_to_float(lex: &mut Lexer<Token>) -> Option<f64> {
-    let slice = lex.slice();
-    let f: f64 = slice.parse().ok()?;
-    Some(f)
-}
-
-fn str_to_int(lex: &mut Lexer<Token>) -> Option<i64> {
-    let slice = lex.slice();
-    let i: i64 = slice.parse().ok()?;
-    Some(i)
-}
-
-#[derive(Logos, Debug, PartialEq)]
-#[logos(extras = LexerInfo)]
-enum Token {
-    #[regex(r#"[\p{L}_][\p{L}\d_]*"#, str_to_id)]
-    Id(Id),
-    #[regex(r#""([^"\\]|\\t|\\u|\\n|\\")*""#, str_to_string)]
-    Str(String),
-    #[regex(r#"[-+]?[0-9]*\.?[0-9]+([eE][-+]?[0-9]+)?"#, str_to_float)]
-    FloatVal(f64),
-    #[regex(r#"[-+]?\d+"#, str_to_int, priority = 2)]
-    IntVal(i64),
+#[derive(Logos, Clone, Debug, PartialEq)]
+pub enum Token<'input> {
+    #[regex(r#"[\p{L}_][\p{L}\d_]*"#)]
+    Id(&'input str),
+    #[regex(r#""([^"\\]|\\t|\\u|\\n|\\")*""#)]
+    Str(&'input str),
+    #[regex(r#"[-+]?[0-9]*\.?[0-9]+([eE][-+]?[0-9]+)?"#)]
+    FloatVal(&'input str),
+    #[regex(r#"[-+]?\d+"#, priority = 2)]
+    IntVal(&'input str),
     #[token("i32", str_to_var_type)]
     #[token("i64", str_to_var_type)]
     #[token("f32", str_to_var_type)]
@@ -222,8 +81,7 @@ enum Token {
     #[token("extends")]
     Extends,
     #[error]
-    #[regex(r"[\n]+", count_newline)]
-    #[regex(r"[ \t\f\r]+", logos::skip)]
+    #[regex(r"[\s\f\r]+", logos::skip)]
     Error,
 }
 
@@ -354,17 +212,19 @@ mod test {
         let lex = Token::lexer(input);
         let tokens: Vec<Token> = lex.into_iter().collect();
 
+        let i64_max_str = i64::MAX.to_string();
+
         assert_eq!(
             tokens,
             vec![
-                Token::IntVal(1),
-                Token::FloatVal(2.0),
-                Token::FloatVal(3.1),
-                Token::FloatVal(4.234),
-                Token::IntVal(5),
-                Token::IntVal(123456789),
-                Token::IntVal(i64::MAX),
-                Token::FloatVal(f64::MAX),
+                Token::IntVal("1"),
+                Token::FloatVal("2.0"),
+                Token::FloatVal("3.1"),
+                Token::FloatVal("4.234"),
+                Token::IntVal("5"),
+                Token::IntVal("123456789"),
+                Token::IntVal(i64_max_str.as_str()),
+                Token::FloatVal("1.7976931348623157E+308"),
             ]
         );
     }
@@ -378,8 +238,8 @@ mod test {
         assert_eq!(
             tokens,
             vec![
-                Token::Str("This is a test string\n".to_string()),
-                Token::Str("This is another test string\t".to_string())
+                Token::Str("\"This is a test string\\n\""),
+                Token::Str("\"This is another test string\\t\"")
             ]
         );
     }
