+++ src/main.rs
@@ -1,3 +1,4 @@
+mod ast;
 mod lexer;
 
 fn main() {
